refactor(categories): rename and reorganize category-related classes and interfaces for improved clarity
