+++ app/src/main/java/com/giraffe/tudeeapp/di/appModule.kt
@@ -7,7 +7,7 @@ import com.giraffe.tudeeapp.data.service.TasksServiceImp
 import com.giraffe.tudeeapp.domain.service.CategoriesService
 import com.giraffe.tudeeapp.domain.service.SplashService
 import com.giraffe.tudeeapp.domain.service.TasksService
-import com.giraffe.tudeeapp.presentation.categories.CategoryViewModel
+import com.giraffe.tudeeapp.presentation.categories.viewmodel.CategoryViewModel
 import com.giraffe.tudeeapp.presentation.categories.tasks_by_category.TasksByCategoryViewModel
 import com.giraffe.tudeeapp.presentation.splash.viewmodel.SplashViewModel
 import org.koin.android.ext.koin.androidApplication
