+++ app/src/main/java/com/giraffe/tudeeapp/presentation/categories/state/CategoriesUiEvent.kt
@@ -1,4 +1,4 @@
-package com.giraffe.tudeeapp.presentation.categories.uiEvent
+package com.giraffe.tudeeapp.presentation.categories.state
 
 sealed class CategoriesUiEvent {
     data class NavigateToTasksByCategoryScreen(val categoryId: Long) : CategoriesUiEvent()
