+++ app/src/main/java/com/giraffe/tudeeapp/presentation/categories/state/CategoriesUiState.kt
@@ -1,9 +1,9 @@
-package com.giraffe.tudeeapp.presentation.categories.uistates
+package com.giraffe.tudeeapp.presentation.categories.state
 
 import com.giraffe.tudeeapp.domain.model.Category
 import com.giraffe.tudeeapp.domain.util.DomainError
 
-data class CategoriesScreenUiState(
+data class CategoriesUiState(
     val categories: List<Category> = emptyList(),
     val isLoading: Boolean = true,
     val error: DomainError? = null,
