+++ app/src/main/java/com/giraffe/tudeeapp/presentation/categories/viewmodel/CategoryViewModel.kt
@@ -1,4 +1,4 @@
-package com.giraffe.tudeeapp.presentation.categories
+package com.giraffe.tudeeapp.presentation.categories.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
@@ -6,8 +6,9 @@ import com.giraffe.tudeeapp.domain.model.Category
 import com.giraffe.tudeeapp.domain.service.CategoriesService
 import com.giraffe.tudeeapp.domain.util.onError
 import com.giraffe.tudeeapp.domain.util.onSuccess
-import com.giraffe.tudeeapp.presentation.categories.uiEvent.CategoriesUiEvent
-import com.giraffe.tudeeapp.presentation.categories.uistates.CategoriesScreenUiState
+import com.giraffe.tudeeapp.presentation.categories.state.CategoriesAction
+import com.giraffe.tudeeapp.presentation.categories.state.CategoriesUiEvent
+import com.giraffe.tudeeapp.presentation.categories.state.CategoriesUiState
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.delay
@@ -20,10 +21,10 @@ import kotlinx.coroutines.launch
 
 class CategoryViewModel(
     private val categoriesService: CategoriesService,
-) : ViewModel(), CategoriesScreenAction {
+) : ViewModel(), CategoriesAction {
 
-    private var _categoriesUiState = MutableStateFlow(CategoriesScreenUiState())
-    val categoriesUiState: StateFlow<CategoriesScreenUiState> = _categoriesUiState.asStateFlow()
+    private var _categoriesUiState = MutableStateFlow(CategoriesUiState())
+    val categoriesUiState: StateFlow<CategoriesUiState> = _categoriesUiState.asStateFlow()
 
     private val _events = Channel<CategoriesUiEvent>()
     val events = _events.receiveAsFlow()
